по итогу разбора домашнего задания вынес все в новый пакет в котором есть интерфейс м класс Dequeue
