+++ src/dz_Lesson_3/queue/Queue.java
@@ -11,8 +11,5 @@ public interface Queue<E> extends ICollection {
 
     E peek();
 
-    E removeRight();
-
-    boolean insertLeft (E value);
 
 }
